+++ flatlaf-core/src/main/java/com/formdev/flatlaf/icons/FlatWindowIconifyIcon.java
@@ -27,11 +27,17 @@ public class FlatWindowIconifyIcon
 	extends FlatWindowAbstractIcon
 {
 	public FlatWindowIconifyIcon() {
+		this( null );
+	}
+
+	/** @since 3.2 */
+	public FlatWindowIconifyIcon( String windowStyle ) {
+		super( windowStyle );
 	}
 
 	@Override
 	protected void paintIconAt1x( Graphics2D g, int x, int y, int width, int height, double scaleFactor ) {
-		int iw = (int) (10 * scaleFactor);
+		int iw = (int) (getSymbolHeight() * scaleFactor);
 		int ih = (int) scaleFactor;
 		int ix = x + ((width - iw) / 2);
 		int iy = y + ((height - ih) / 2);